Introducing Search Tree CutOff
* Due to the exploding space & time complexity of the problem, it was decided we were better off pruning the siblings in each level of the DFS expansion, as well as to stop in some cases the depth expansion.
After playing around with the parameters the following estimates did not too bad:
=> In the large (10x10) game: 40% of the time we stop the depth expansion, 40% of the children in each level are not even considered.
=> In the small (6x6) game: only 30% of the time we stop the depth expansion, and only 10% of the children in each level are ignored.

In a game like amazons the complexity is simply too great for the computer to handle deep levels with many children.

Either way, these parameters are up for further experimentation, but I'm content with the results for this project. They are: ShouldCutOffSiblings and ShouldCutOffDepth.